refactor: Reuse GetParticipant from UpdateParticipantData
