+++ application/CohortManager/src/Functions/screeningDataServices/createParticipant/Program.cs
@@ -20,6 +20,7 @@ var host = new HostBuilder()
             return conn;
         });
         services.AddSingleton<ICreateParticipantData, CreateParticipantData>();
+        services.AddSingleton<IUpdateParticipantData, UpdateParticipantData>();
         services.AddSingleton<ICreateResponse, CreateResponse>();
         services.AddSingleton<IDatabaseHelper, DatabaseHelper>();
     })
